% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{evaluate_svg_detection}
\alias{evaluate_svg_detection}
\title{Evaluate spatially variable gene detection}
\usage{
evaluate_svg_detection(truth, pred)
}
\arguments{
\item{truth}{Binary vector of true SVG labels.}

\item{pred}{Binary vector of predicted SVG labels.}
}
\value{
A list containing a one-row summary of detection metrics and a
two-by-two confusion matrix.
}
\description{
Evaluate spatially variable gene detection
}
