% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{select_svg_bfdr}
\alias{select_svg_bfdr}
\title{Select spatially variable genes using Bayesian FDR control}
\usage{
select_svg_bfdr(p0g, alpha = 0.05)
}
\arguments{
\item{p0g}{Numeric vector containing posterior probabilities of no spatial
effect.}

\item{alpha}{Target Bayesian false discovery rate.}
}
\value{
A list containing the binary selection vector, selected gene
indices, ordering information, cumulative Bayesian FDR values, and the
selected cutoff.
}
\description{
Select spatially variable genes using Bayesian FDR control
}
