% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{compute_p0g}
\alias{compute_p0g}
\title{Compute posterior probabilities of no spatial effect}
\usage{
compute_p0g(fit_res, tol = 0)
}
\arguments{
\item{fit_res}{An object returned by \code{fit_all_genes()}.}

\item{tol}{Nonnegative numerical tolerance used to define a zero
coefficient.}
}
\value{
A numeric vector containing one posterior null probability per gene.
}
\description{
For each gene, calculates the posterior proportion of MCMC samples in
which all spatial-basis coefficients are zero, or within a specified
tolerance of zero.
}
