% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\docType{package}
\name{GeoSVG3D-package}
\alias{GeoSVG3D}
\alias{GeoSVG3D-package}
\title{GeoSVG3D package}
\description{
Geometry-aware Bayesian detection of spatially variable genes in
three-dimensional spatial transcriptomics data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Yuanzihui0810/GeoSVG3D}
}

}
\author{
\strong{Maintainer}: Zihui Yuan \email{yuanzihui0810@gmail.com}

}
\keyword{internal}
